+++ packages/adapter-evm/src/chains/common.ts
@@ -1,4 +1,4 @@
-import { Adapter, AdapterFactory, WarpChainEnv, WarpChainInfo, WarpClientConfig } from '@vleap/warps'
+import { Adapter, AdapterFactory, WarpChain, WarpChainEnv, WarpChainInfo, WarpClientConfig } from '@vleap/warps'
 import { WarpEvmDataLoader } from '../WarpEvmDataLoader'
 import { WarpEvmExecutor } from '../WarpEvmExecutor'
 import { WarpEvmExplorer } from '../WarpEvmExplorer'
@@ -6,17 +6,12 @@ import { WarpEvmResults } from '../WarpEvmResults'
 import { WarpEvmSerializer } from '../WarpEvmSerializer'
 import { WarpEvmWallet } from '../WarpEvmWallet'
 
-export const createEvmAdapter = (
-  chainName: string,
-  chainPrefix: string,
-  chainInfos: Record<WarpChainEnv, WarpChainInfo>
-): AdapterFactory => {
+export const createEvmAdapter = (chainName: WarpChain, chainInfos: Record<WarpChainEnv, WarpChainInfo>): AdapterFactory => {
   return (config: WarpClientConfig, fallback?: Adapter) => {
     if (!fallback) throw new Error(`${chainName} adapter requires a fallback adapter`)
 
     return {
       chainInfo: chainInfos[config.env],
-      prefix: chainPrefix,
       builder: () => fallback.builder(),
       executor: new WarpEvmExecutor(config, chainInfos[config.env]),
       results: new WarpEvmResults(config, chainInfos[config.env]),