+++ packages/adapter-multiversx/src/WarpMultiversxAbiBuilder.ts
@@ -1,6 +1,7 @@
 import { AbiRegistry, Address, TransactionOnNetwork, TransactionsFactoryConfig, TransferTransactionsFactory } from '@multiversx/sdk-core'
 import {
   AdapterWarpAbiBuilder,
+  createWarpIdentifier,
   getLatestProtocolIdentifier,
   getWarpWalletAddressFromConfig,
   WarpAbi,
@@ -65,6 +66,7 @@ export class WarpMultiversxAbiBuilder implements AdapterWarpAbiBuilder {
 
     abi.meta = {
       chain: this.chain.name,
+      identifier: createWarpIdentifier(this.chain.name, 'hash', tx.hash),
       hash: tx.hash,
       creator: tx.sender.toBech32(),
       createdAt: new Date(tx.timestamp * 1000).toISOString(),