+++ packages/adapter-multiversx/src/WarpMultiversxBrandBuilder.ts
@@ -1,5 +1,13 @@
 import { Address, Transaction, TransactionOnNetwork, TransactionsFactoryConfig, TransferTransactionsFactory } from '@multiversx/sdk-core'
-import { getWarpWalletAddressFromConfig, WarpBrand, WarpBrandBuilder, WarpChainInfo, WarpClientConfig, WarpLogger } from '@vleap/warps'
+import {
+  createWarpIdentifier,
+  getWarpWalletAddressFromConfig,
+  WarpBrand,
+  WarpBrandBuilder,
+  WarpChainInfo,
+  WarpClientConfig,
+  WarpLogger,
+} from '@vleap/warps'
 import { Buffer } from 'buffer'
 import { getMultiversxEntrypoint } from './helpers/general'
 
@@ -33,6 +41,7 @@ export class WarpMultiversxBrandBuilder {
 
     brand.meta = {
       chain: this.chain.name,
+      identifier: createWarpIdentifier(this.chain.name, 'hash', tx.hash),
       hash: tx.hash,
       creator: tx.sender.toBech32(),
       createdAt: new Date(tx.timestamp * 1000).toISOString(),