+++ packages/adapter-multiversx/src/WarpMultiversxBuilder.ts
@@ -1,6 +1,7 @@
 import { Address, TransactionOnNetwork, TransactionsFactoryConfig, TransferTransactionsFactory } from '@multiversx/sdk-core'
 import {
   AdapterWarpBuilder,
+  createWarpIdentifier,
   getWarpWalletAddressFromConfig,
   Warp,
   WarpBuilder,
@@ -51,6 +52,7 @@ export class WarpMultiversxBuilder extends WarpBuilder implements AdapterWarpBui
 
     warp.meta = {
       chain: this.chain.name,
+      identifier: createWarpIdentifier(this.chain.name, 'hash', tx.hash),
       hash: tx.hash,
       creator: tx.sender.toBech32(),
       createdAt: new Date(tx.timestamp * 1000).toISOString(),